+++ client/src/components/AddTask.jsx
@@ -1,7 +1,6 @@
 import {HiPlusCircle} from 'react-icons/hi';
 
 const AddTask = ({addTask, containerId}) => {
-    console.log(containerId);
     return (
         <button
             className='flex gap-2 items-center border-columnBackgroundColor border-2 rounded-md p-4 border-x-columnBackgroundColor hover:bg-mainBackgroundColor hover:bg-warningLight'
