--- client/src/components/data.jsx
@@ -1,191 +1,0 @@
-const colums = [
-    {
-        title: 'Backlog',
-        tasks: [
-            {
-                taskName: 'Task 1!',
-                assignedUserIds: [],
-                priority: 1,
-                tags: [],
-                taskDetail: 'Add more in-depth task information here!',
-                _id: '64df48c0a0d786ac685cb36f',
-                created_at: '2023-08-18T10:32:32.170Z',
-                comments: [],
-            },
-            {
-                taskName: 'Task 2',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cd8',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Task 3',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cd9',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Task 4',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cd10',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-        ],
-        _id: '64df48c0a0d786ac685cb36e',
-    },
-
-    {
-        title: 'Todo',
-        tasks: [
-            {
-                taskName: 'Add your tasks here!',
-                assignedUserIds: [],
-                priority: 1,
-                tags: [],
-                taskDetail: 'Add more in-depth task information here!',
-                _id: '64df48c0a0d786ac685cb3a1',
-                created_at: '2023-08-18T10:32:32.170Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26ca2',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26ca3',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cda4',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-        ],
-        _id: '64df48c0a0d786ac685cb370',
-    },
-
-    {
-        title: 'In Progress',
-        tasks: [
-            {
-                taskName: 'Add your tasks here!',
-                assignedUserIds: [],
-                priority: 1,
-                tags: [],
-                taskDetail: 'Add more in-depth task information here!',
-                _id: '64df48c0a0d786ac685cb3b1',
-                created_at: '2023-08-18T10:32:32.170Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cb2',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cb3',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cdb4',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-        ],
-        _id: '64df48c0a0d786ac685cb371',
-    },
-
-    {
-        title: 'Done',
-        tasks: [
-            {
-                taskName: 'Add your tasks here!',
-                assignedUserIds: [],
-                priority: 1,
-                tags: [],
-                taskDetail: 'Add more in-depth task information here!',
-                _id: '64df48c0a0d786ac685cb36c1',
-                created_at: '2023-08-18T10:32:32.170Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cc2',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cc3',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-            {
-                taskName: 'Backlog Task 8-12',
-                assignedUserIds: [],
-                priority: 2,
-                tags: [],
-                taskDetail: 'This is a test from postman',
-                _id: '64e0ad59f326220bada26cdc4',
-                created_at: '2023-08-19T11:54:01.192Z',
-                comments: [],
-            },
-        ],
-        _id: '64df48c0a0d786ac685cb372',
-    },
-];
-
-export default colums;
