+++ client/src/features/Workspace.jsx
@@ -1,10 +1,8 @@
 import Header from '../components/Header';
-// import KanbanBoard from '../components/KanbanBoard';
 import WorkspaceHeader from './WorkspaceHeader';
-import {useSelectedBoardContext} from '../contexts/BoardContext/boardContext';
-
 import WorkspaceSlideOver from './WorkspaceSlideOver';
-import KANBAN from '../components/KANBAN';
+import KanbanBoard from './kanbanBoard/KanbanBoard';
+import {useSelectedBoardContext} from '../contexts/BoardContext/boardContext';
 
 const Workspace = () => {
     const {selectedBoard} = useSelectedBoardContext();
@@ -16,7 +14,7 @@ const Workspace = () => {
         <div>
             <Header />
             <WorkspaceHeader boardInfo={boardInfo} />
-            <KANBAN boardInfo={boardInfo} />
+            <KanbanBoard boardInfo={boardInfo} />
         </div>
     );
 };
